+++ srcs/ddos/get_flood.c
@@ -12,11 +12,11 @@
 extern int g_num_threads;
 extern int g_packet_size;
 // session counting
-__u64 g_get_num_total;
-__u64 g_get_num_generated_in_sec;
+uint32_t g_get_num_total;
+uint32_t g_get_num_generated_in_sec;
 // from main()
 InputArguments g_get_input;
-__u32 g_get_request_per_sec;
+uint32_t g_get_request_per_sec;
 // for masking next ip address
 MaskingArguments g_get_now;
 // thread
